Extract Formatting Rules
